Use time function from the windows_linux folder
