+++ Platformio/hardware/windows_linux/util_hal_windows_linux.h
@@ -1,0 +1,1 @@
+long long current_timestamp_hal_windowsLinux();
